Final version of the scripts

+++ src/phraseClusteringbasedShallowSummaryKmedoid-New-Malformed-LexRank.py
@@ -9,28 +9,13 @@ import NLTKWrapper
 import SennaParser
 import porter
 import phraseClusteringKmedoid
+import postProcess
 
 stopwords = [line.lower().strip() for line in fio.readfile("../../ROUGE-1.5.5/data/smart_common_words.txt")]
 #print "stopwords:", len(stopwords)
 
 stopwords = stopwords + ['.', '?', '-', ',', '[', ']', '-', ';', '\'', '"', '+', '&', '!', '/', '>', '<', ')', '(', '#', '=']
-
-def getTopRankPhrase(NPs, clusterids, cid, lexdict, sources):
-    #get cluster NP, and scores
-    dict = {}
-    
-    s = []
-    
-    for NP, id, source in zip(NPs, clusterids, sources):
-        if int(id) == cid:
-            dict[NP] = lexdict[NP.lower()]
-            s.append(source)
-    
-    keys = sorted(dict, key=dict.get, reverse =True)
-    
-    source = set(s)
-    return keys[0], source
-                            
+                                
 def getShallowSummary(excelfile, folder, sennadatadir, clusterdir, K=30, method=None, ratio=None, np=None, lex='lexrank'):
     #K is the number of words per points
     header = ['ID', 'Gender', 'Point of Interest', 'Muddiest Point', 'Learning Point']
@@ -64,9 +49,22 @@ def getShallowSummary(excelfile, folder, sennadatadir, clusterdir, K=30, method=
             output = clusterdir + str(week) +'/' + type + ".cluster.kmedoids." + str(ratio) + "." +method + '.' + np
             weightfile = clusterdir + str(week)+ '/' + type + '.' + np + '.' + method
             if not fio.isExist(output):
+            #if True:
                 phraseClusteringKmedoid.getPhraseClusterAll(sennafile, weightfile, output, ratio, MalformedFlilter=True, source=ids, np=np)
             
             NPCandidates, sources = phraseClusteringKmedoid.getNPs(sennafile, MalformedFlilter=True, source=ids, np=np)
+            
+            body = []
+            for NP, s in zip(NPCandidates, sources):
+                row = []
+                row.append(NP)
+                row.append(s)
+                body.append(row)
+            
+            sourcelist = clusterdir + str(week) + "/" + type + '.' + np + '.source.txt'
+            print sourcelist
+            
+            fio.writeMatrix(sourcelist,body,None)
             
             #write the sources
             sourcedict = {}
@@ -90,27 +88,55 @@ def getShallowSummary(excelfile, folder, sennadatadir, clusterdir, K=30, method=
             Summary = []
             
             #sort the clusters according to the number of phrases
-            dict = defaultdict(float)
-            for row in body: 
-                dict[int(row[1])] = dict[int(row[1])] + 1
+#             dict = defaultdict(float)
+#             for row in body: 
+#                 dict[int(row[1])] = dict[int(row[1])] + 1
+#             
+#             keys = sorted(dict, key=dict.get, reverse =True)
             
-            keys = sorted(dict, key=dict.get, reverse =True)
+            #sort the clusters according to the number of students
+            
+            keys = postProcess.RankCluster(NPs, lexdict, clusterids, sources)
             
             sumarysource = []
             
+            #sort by number of students
             total_word = 0
             word_count = 0
             for key in keys:
                 #phrase = NPs[key]
-                phrase, source = getTopRankPhrase(NPs, clusterids, key, lexdict, sources)
+                phrase, source = postProcess.getTopRankPhrase(NPs, clusterids, int(key), lexdict, sources)
                 if phrase in Summary: continue
-                
+                  
                 word_count = len(phrase.split())
                 total_word = total_word + word_count
-                #if total_word <= K:
-                if len(Summary) + 1 <= K:
+                if total_word <= K:
+                #if len(Summary) + 1 <= K:
                     Summary.append(phrase)
                     sumarysource.append(",".join(source))
+            
+#             added_cluster = []
+#             total_word = 0
+#             word_count = 0
+#             for phrase in sorted(lexdict, key=lexdict.get, reverse=True):
+#                 if phrase not in NPs: continue
+#                 key = NPs.index(phrase)
+#                 if key == -1: continue
+#                  
+#                 key = clusterids[key]
+#                 if key in added_cluster: continue
+#                 added_cluster.append(key)
+#                  
+#                 if phrase in Summary: continue
+#                 word_count = len(phrase.split())
+#                 total_word = total_word + word_count
+#                  
+# #                _, source = getTopRankPhrase(NPs, clusterids, key, lexdict, sources)
+#                  
+#                 #if total_word <= K:
+#                 if len(Summary) + 1 <= K:
+#                     Summary.append(phrase)
+#                     sumarysource.append(",".join(source))
             
             fio.savelist(Summary, filename)
             fio.savelist(sumarysource, filename + ".source")
@@ -130,8 +156,7 @@ def GetLexRankScore(datadir, np, outputdir):
             DID = str(week) + '_' + type
             
             phrases = []
-            scores = []
-    
+            
             #read Docsent
             path = datadir + str(week)+ '/'
             path = path + type + '/'
@@ -145,22 +170,35 @@ def GetLexRankScore(datadir, np, outputdir):
                 phrases.append(child.text)
             
             #read feature
-            path = datadir + str(week)+ '/'
-            path = path + type + '/'
-            path = path + 'feature/'
-            filename = path + type + '.LexRank.sentfeature'
-            
-            tree = ET.parse(filename)
-            root = tree.getroot()
-            
-            for child in root:
-                feature = child[0]
-                #print feature.tag, feature.attrib, feature.attrib['V']
-                #print child.tag, child.attrib
-                scores.append(feature.attrib['V'])
+            Fscores = []
+            for feature in ['LexRank']: #'Position', 
+                scores = []
+                path = datadir + str(week)+ '/'
+                path = path + type + '/'
+                path = path + 'feature/'
+                filename = path + type + '.'+feature+'.sentfeature'
                 
-            #write
-            assert(len(phrases) == len(scores))
+                tree = ET.parse(filename)
+                root = tree.getroot()
+                
+                for child in root:
+                    feature = child[0]
+                    #print feature.tag, feature.attrib, feature.attrib['V']
+                    #print child.tag, child.attrib
+                    scores.append(feature.attrib['V'])
+                    
+                #write
+                assert(len(phrases) == len(scores))
+                
+                Fscores.append(scores)
+            
+            #average the score
+            scores = []
+            for i in range(len(phrases)):
+                ave = 0
+                for s in Fscores:
+                    ave = ave + float(s[i])
+                scores.append(ave/len(Fscores))
             
             dict = {}
             for phrase, score in zip(phrases, scores):
@@ -183,25 +221,31 @@ if __name__ == '__main__':
     excelfile = "../data/2011Spring.xls"
     
     sennadatadir = "../data/senna/"
-    clusterdir = "../data/np/"
-    
-    for np in ['chunk', 'syntax']:
-        datadir = "../../mead/data/PhraseMeadLexRank_"+np+"/"
-        GetLexRankScore(datadir, np, clusterdir)
+
+    #for i in range(1, 1000):
+    if True:
+        i = "511"
+        #clusterdir = "../data/np/"
+        clusterdir = "../data/np"+str(i)+"/"
         
-    for ratio in ["sqrt"]:
-    #for ratio in ["sqrt", 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]:
-        #for method in ['optimumComparerLSATasa']: #'bleuComparer', 'cmComparer', 'lsaComparer',
-        #for method in ['dependencyComparerWnLeskTanim']:
-        for method in ['npsoft', 'optimumComparerLSATasa']:
-        #for method in ['optimumComparerLSATasa']:
-            for np in ['chunk', 'syntax']:
-            #for np in ['syntax']:
-                #for lex in ['lexrank', 'lexrankmax']:
-                for lex in ['lexrankmax']:
-                    datadir = "../../mead/data/C4_Clustering_"+lex+"_"+str(ratio)+"_"+method+"_"+np+"/"   
-                    fio.deleteFolder(datadir)
-                    ShallowSummary(excelfile, datadir, sennadatadir, clusterdir, K=4, method=method, ratio=ratio, np=np, lex=lex)
-            
+        for np in ['syntax']:
+            datadir = "../../mead/data/PhraseMeadLexRank_"+np+"/"
+            GetLexRankScore(datadir, np, clusterdir)
+        
+        for ratio in ["sqrt"]:
+        #for ratio in [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]:
+            #for method in ['optimumComparerLSATasa']: #'bleuComparer', 'cmComparer', 'lsaComparer',
+            #for method in ['dependencyComparerWnLeskTanim']:
+            #for method in ['npsoft', 'optimumComparerLSATasa']:
+            for method in ['optimumComparerLSATasa']:
+                for np in ['syntax']:
+                #for np in ['syntax']:
+                    #for lex in ['lexrank', 'lexrankmax']:
+                    for lex in ['lexrankmax']:
+                        #datadir = "../../mead/data/C4_Clustering_"+lex+"_"+str(ratio)+"_"+method+"_"+np+"/"   
+                        datadir = "../../mead/data/C30_ClusterARank"+str(i)+"/"
+                        fio.deleteFolder(datadir)
+                        ShallowSummary(excelfile, datadir, sennadatadir, clusterdir, K=30, method=method, ratio=ratio, np=np, lex=lex)
+                
     print "done"
